Decompose execution.py
dispatcher.py:
- `_build_sublattice_graph
- `_dispatch_sync_sublattice`
- `_get_abstract_task_inputs`
- `_post_process`
- `_handle_completed_node`
- `_handle_failed_node`
- `_handle_cancelled_node`
- `_initialize_deps_and_queue`
- `_postprocess_workflow`
- `_submit_task`
- `_run_planned_workflow`
- `_plan_workflow`
- `run_workflow`
- `cancel_workflow`

result.py
- `generate_node_result`
- `_update_node_result`
- `initialize_result_object`
- `get_unique_id`

runner.py
- `_get_task_input_values`
- `_get_task_inputs`
- `_run_abstract_task`
- `_run_task`
- `_gather_deps`
- `_run_task_and_update
