+++ covalent_dispatcher/_core/__init__.py
@@ -18,4 +18,5 @@
 #
 # Relief from the License may be granted by purchasing a commercial license.
 
-from .execution import cancel_workflow, initialize_result_object, run_workflow
+from .dispatcher import cancel_workflow, run_workflow
+from .result import initialize_result_object
